+++ src/presentation/controllers/login/login.spec.ts
@@ -1,8 +1,8 @@
 import { Authentication, EmailValidator, HttpRequest } from './login-protocols'
 import { InvalidParamError } from '../../errors'
-import { badRequest, ok, serverError, unauthorized } from '../../helpers/http-helper'
+import { badRequest, ok, serverError, unauthorized } from '../../helpers/http/http-helper'
 import { LoginController } from './login'
-import { Validation } from '../../helpers/validators/validation'
+import { Validation } from '../../protocols/validation'
 
 interface SutInterface {
   sut: LoginController
