+++ src/presentation/controllers/login/login.ts
@@ -1,6 +1,6 @@
 import { Authentication, Controller, HttpRequest, HttpResponse } from './login-protocols'
-import { badRequest, ok, serverError, unauthorized } from '../../helpers/http-helper'
-import { Validation } from '../../helpers/validators/validation'
+import { badRequest, ok, serverError, unauthorized } from '../../helpers/http/http-helper'
+import { Validation } from '../../protocols/validation'
 
 export class LoginController implements Controller {
   private readonly validation: Validation
